// src/graphql/generated/types.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Upload: any;
};
// ...
export type ReplyMessageInput = {
  chat_type: Scalars["String"];
  cid: Scalars["String"];
  created_at: Scalars["Float"];
  fontStyle: Scalars["String"];
  message: Scalars["String"];
  name: Scalars["String"];
  sender: Scalars["String"];
  url: Scalars["String"];
};
// ...
export type SendChatDataInput = {
  _id?: InputMaybe<Scalars["String"]>;
  duration?: InputMaybe<Scalars["Float"]>;
  fileURL?: InputMaybe<Scalars["String"]>;
  fontStyle?: InputMaybe<Scalars["String"]>;
  id_local?: InputMaybe<Scalars["String"]>;
  isForwarded?: InputMaybe<Scalars["Boolean"]>;
  message: Scalars["String"];
  roomId: Scalars["String"];
  thumbnail?: InputMaybe<Scalars["String"]>;
  type: Scalars["String"];
};
// ...
export type UdpateChatInput = {
  data: SendChatDataInput;
  isSent: Scalars["Boolean"];
  reply_msg?: InputMaybe<ReplyMessageInput>;
};